+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.20.51
 
+* `Versions`:
+    * `Android Fragment`: `1.7.0` -> `1.7.1`
 * `Pagination`:
     * Add `Pagination.nextPageIfTrue` and `Pagination.thisPageIftrue` extensions to get the page according to condition
     pass status